freetype: install test exe and move all tests to src/main.zig
