+++ freetype/src/Face.zig
@@ -3,7 +3,7 @@ const utils = @import("utils");
 const c = @import("c");
 const intToError = @import("error.zig").intToError;
 const Error = @import("error.zig").Error;
-const GlyphSlot = @import("freetype.zig").GlyphSlot;
+const GlyphSlot = @import("GlyphSlot.zig");
 const LoadFlags = @import("freetype.zig").LoadFlags;
 const FaceFlags = @import("freetype.zig").FaceFlags;
 const StyleFlags = @import("freetype.zig").StyleFlags;
