+++ freetype/src/GlyphSlot.zig
@@ -2,11 +2,11 @@ const c = @import("c");
 const intToError = @import("error.zig").intToError;
 const Error = @import("error.zig").Error;
 const Glyph = @import("glyph.zig").Glyph;
-const Library = @import("freetype.zig").Library;
-const Face = @import("freetype.zig").Face;
+const Library = @import("Library.zig");
+const Face = @import("Face.zig");
 const RenderMode = @import("freetype.zig").RenderMode;
 const Matrix = @import("types.zig").Matrix;
-const Outline = @import("image.zig").Outline;
+const Outline = @import("Outline.zig");
 const GlyphFormat = @import("image.zig").GlyphFormat;
 const Vector = @import("image.zig").Vector;
 const GlyphMetrics = @import("image.zig").GlyphMetrics;
