+++ freetype/src/glyph.zig
@@ -3,12 +3,12 @@ const c = @import("c");
 const intToError = @import("error.zig").intToError;
 const Error = @import("error.zig").Error;
 const Stroker = @import("stroke.zig").Stroker;
-const Library = @import("freetype.zig").Library;
+const Library = @import("Library.zig");
 const RenderMode = @import("freetype.zig").RenderMode;
 const SizeMetrics = @import("freetype.zig").SizeMetrics;
 const Matrix = @import("types.zig").Matrix;
 const BBox = @import("types.zig").BBox;
-const Outline = @import("image.zig").Outline;
+const Outline = @import("Outline.zig");
 const GlyphFormat = @import("image.zig").GlyphFormat;
 const Vector = @import("image.zig").Vector;
 const Bitmap = @import("image.zig").Bitmap;
