--- freetype/test/ft/ft2build.h
@@ -1,9 +1,0 @@
-#ifndef FT2_BUILD_MACH_H_
-#define FT2_BUILD_MACH_H_
-
-#define FT_CONFIG_OPTIONS_H <mach_ftoption.h>
-#define FT_CONFIG_MODULES_H <mach_ftmodule.h>
-
-#include <freetype/config/ftheader.h>
-
-#endif
